requirements.txt
# Core dependencies
python-dotenv==1.0.1
pydantic==2.10.3

# LLM
llama-cpp-python==0.3.9

# Database - используем самые новые версии
asyncpg>=0.30.0
sqlalchemy>=2.0.36
alembic>=1.14.0

# Telegram Bot
python-telegram-bot==21.7

# Communication tools
aiosmtplib==3.0.1

# LDAP (optional)
ldap3==2.9.1

# Utilities
requests==2.32.3
python-dateutil==2.8.2
fastapi~=0.115.12